# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: TSITA GEST × BSCA Bank — API
  version: v1
  description: |
    API REST de TSITA GEST, outil de gestion des réclamations clients de BSCA Bank
    (contrat de référence : `docs/ARCHITECTURE.md` §9). **Toutes les données sont fictives**
    (démonstration) : aucun chiffre ni client réel BSCA.

    ## Authentification (Laravel Sanctum, mode SPA par cookie)
    1. Appeler `GET /sanctum/csrf-cookie` (hors préfixe `/api/v1`, à la racine du serveur) :
       le serveur pose le cookie `XSRF-TOKEN`.
    2. Envoyer, sur **chaque** requête modifiante (POST/PATCH/DELETE), l'en-tête `X-XSRF-TOKEN`
       contenant la valeur (décodée URL) du cookie `XSRF-TOKEN`, avec les cookies (`withCredentials`).
    3. `POST /auth/login` ouvre la session : cookie httpOnly `tsita-gest-session` (SameSite=Lax),
       régénéré à la connexion, expiration après 30 min d'inactivité.

    Les routes `/public/*`, `/auth/login`, `/auth/forgot-password` et `/auth/reset-password`
    ne nécessitent pas de session (le jeton CSRF reste requis pour les POST issus du navigateur).

    ## Conventions
    - Ressources enveloppées `{ "data": … }` ; listes paginées
      `{ "data": [...], "meta": { "current_page", "last_page", "per_page", "total" } }`.
    - Dates-heures : ISO 8601 **avec décalage `+01:00`** (fuseau `Africa/Brazzaville`, sans heure
      d'été ; stockage UTC). Dates seules : `AAAA-MM-JJ` (jour local).
...
      (agence / entité / propriété du dossier), appliqués côté serveur.
...

...
servers:
...
security:
  - sessionCookie: []
    xsrfToken: []
...
paths:
...
  /deadlines:
    get:
      tags: [Pilotage]
      operationId: deadlines
      summary: Files d'échéance (paginées) + synthèse + règles actives
      description: |
        `a_echeance` (défaut) : dossiers ouverts non en retard, par échéance croissante ;
        `en_retard` : ouverts dont l'échéance finale est dépassée ; `clos_en_retard` : répondus après l'échéance.
        Les doublons sont exclus.
      x-permissions: [deadlines.view]
      parameters:
        - $ref: '#/components/parameters/Queue'
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - $ref: '#/components/parameters/AsOf'
        - $ref: '#/components/parameters/AgencyId'
        - $ref: '#/components/parameters/EntityId'
        - $ref: '#/components/parameters/CategoryId'
        - $ref: '#/components/parameters/ProductId'
        - $ref: '#/components/parameters/Channel'
        - $ref: '#/components/parameters/Status'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/DeadlineFlagFilter'
        - $ref: '#/components/parameters/OwnerId'
        - $ref: '#/components/parameters/Mine'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Page de la file demandée.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeadlinesPage' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '429': { $ref: '#/components/responses/TooManyApi' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: tsita-gest-session
      description: Cookie de session Laravel (httpOnly, SameSite=Lax), posé par `POST /auth/login`. Nom dérivé de APP_NAME (« TSITA GEST » → `tsita-gest-session`).
    xsrfToken:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
      description: Valeur du cookie `XSRF-TOKEN` obtenu via `GET /sanctum/csrf-cookie` ; requis sur les requêtes modifiantes.
  parameters:
    From:
      name: from
      in: query
      description: Début de période (jour local, inclus).
      schema: { type: string, format: date }
    To:
      name: to
      in: query
      description: Fin de période (jour local, inclus) ; ≥ `from`.
      schema: { type: string, format: date }
    AsOf:
      name: as_of
      in: query
      description: Date de calcul (défaut aujourd'hui, Africa/Brazzaville).
      schema: { type: string, format: date }
    AgencyId:
      name: agency_id
      in: query
      description: Agence de réception.
      schema: { type: integer, minimum: 1 }
    EntityId:
      name: entity_id
      in: query
      description: Entité de traitement.
      schema: { type: integer, minimum: 1 }
    CategoryId:
      name: category_id
      in: query
      schema: { type: integer, minimum: 1 }
    ProductId:
      name: product_id
      in: query
      schema: { type: integer, minimum: 1 }
    Channel:
      name: channel
      in: query
      description: Code de canal existant (ex. `agence`, `email`, `portail`).
      schema: { type: string, maxLength: 30 }
    Status:
      name: status
      in: query
      schema: { $ref: '#/components/schemas/ComplaintStatus' }
    Search:
      name: search
      in: query
      description: Recherche sur la référence, l'objet ou le nom du client.
      schema: { type: string, maxLength: 100 }
    DeadlineFlagFilter:
      name: deadline_flag
      in: query
      schema: { $ref: '#/components/schemas/DeadlineFlag' }
    OwnerId:
      name: owner_id
      in: query
      schema: { type: integer, minimum: 1 }
    Mine:
      name: mine
      in: query
      description: '`1` = uniquement mes dossiers (propriétaire ou suppléant).'
      schema: { type: boolean }
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PerPage:
      name: per_page
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    Queue:
      name: queue
      in: query
      description: File d'échéance.
      schema:
        type: string
        enum: [a_echeance, en_retard, clos_en_retard]
        default: a_echeance
...
  responses:
    Unauthenticated:
      description: Non authentifié, ou compte inactif/verrouillé (session invalidée).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { message: Non authentifié. }
    Forbidden:
      description: Permission ou périmètre insuffisant (ou rôle non admis sur la zone client/staff).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { message: Action non autorisée. }
...
    ValidationFailed:
      description: Données invalides ou règle métier non respectée.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationError' }
...
    TooManyApi:
      description: Limite de l'API authentifiée atteinte (180/min par utilisateur).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      description: Erreur simple.
      required: [message]
      properties:
        message: { type: string }
    ValidationError:
      type: object
      description: Erreur de validation ou de règle métier (422). Clé `general` si l'erreur ne concerne pas un champ.
      required: [message, errors]
      properties:
        message: { type: string, example: Le champ reason est obligatoire. }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
          example: { reason: [Le champ reason est obligatoire.] }
...
    PaginationMeta:
      type: object
      description: Métadonnées de pagination.
      required: [current_page, last_page, per_page, total]
      properties:
        current_page: { type: integer }
        last_page: { type: integer }
        per_page: { type: integer }
        total: { type: integer }

    # ---------------------------------------------------------- Types de base
...
    NullableDateTime:
      type: [string, 'null']
      format: date-time
      description: ISO 8601 avec décalage +01:00, ou null.
      example: '2026-04-14T09:30:00+01:00'
    Reference:
      type: string
      pattern: '^TG-BSCA-\d{4}-\d{6}$'
      example: TG-BSCA-2026-000123
...
    IdName:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
    IdLabel:
      type: object
      required: [id, label]
      properties:
        id: { type: integer }
        label: { type: string }
    CodeLabel:
      type: object
      required: [code, label]
      properties:
        code: { type: string }
        label: { type: string }
...
    ComplaintStatus:
      type: string
      description: |
        Statut interne. brouillon (Brouillon), recu (Reçu), a_qualifier (À qualifier), affecte (Affecté),
        en_investigation (En investigation), attente_information (Attente d'information),
        solution_proposee (Solution proposée), a_valider (À valider), reponse_envoyee (Réponse envoyée),
        cloture (Clôturé), reouvert (Réouvert).
      enum: [brouillon, recu, a_qualifier, affecte, en_investigation, attente_information, solution_proposee, a_valider, reponse_envoyee, cloture, reouvert]
...
    Decision:
      type: string
      description: fondee, partiellement_fondee, non_fondee, irrecevable_motivee.
      enum: [fondee, partiellement_fondee, non_fondee, irrecevable_motivee]
    Priority:
      type: string
      enum: [basse, normale, haute, critique]
...
    DeadlineFlag:
      type: string
      description: ok (Dans les délais), a_risque, en_retard, clos_en_retard, sans_regle (Règle à valider).
      enum: [ok, a_risque, en_retard, clos_en_retard, sans_regle]
    DeadlineKind:
      type: string
      enum: [accuse, reponse_finale, prealerte, controle]
...
    DeadlineUnit:
      type: string
      description: calendar (jours calendaires) ou business (jours ouvrés).
      enum: [calendar, business]
...
    RuleSourceType:
      type: string
      enum: [juridique, interne, demonstration]
    RuleStatus:
      type: string
      enum: [brouillon, a_valider, valide, retire]
...
    ComplaintListItem:
      type: object
      description: Élément de liste (ComplaintListItemResource). `customer_name` masqué (« J. M*** ») pour la direction.
      required: [id, reference, subject, customer_name, channel, receiving_agency, processing_entity, category, status, status_label, decision, decision_label, priority, received_at, due_at, deadline_flag, deadline_flag_label, owner]
      properties:
        id: { type: integer }
        reference: { $ref: '#/components/schemas/Reference' }
        subject: { type: string }
        customer_name: { type: [string, 'null'] }
        channel:
          oneOf:
            - $ref: '#/components/schemas/CodeLabel'
            - type: 'null'
        receiving_agency:
          oneOf:
            - $ref: '#/components/schemas/IdName'
            - type: 'null'
        processing_entity:
          oneOf:
            - $ref: '#/components/schemas/IdName'
            - type: 'null'
        category:
          oneOf:
            - $ref: '#/components/schemas/IdLabel'
            - type: 'null'
        status: { $ref: '#/components/schemas/ComplaintStatus' }
        status_label: { type: string }
        decision:
          oneOf:
            - $ref: '#/components/schemas/Decision'
            - type: 'null'
        decision_label: { type: [string, 'null'] }
        priority: { $ref: '#/components/schemas/Priority' }
        received_at: { $ref: '#/components/schemas/NullableDateTime' }
        due_at:
          $ref: '#/components/schemas/NullableDateTime'
          description: Échéance de réponse finale.
        deadline_flag: { $ref: '#/components/schemas/DeadlineFlag' }
        deadline_flag_label: { type: string }
        owner:
          oneOf:
            - $ref: '#/components/schemas/IdName'
            - type: 'null'
...
    DeadlineQueueItem:
      description: Élément de file d'échéance = ComplaintListItem + dates d'échéance.
      allOf:
        - $ref: '#/components/schemas/ComplaintListItem'
        - type: object
          required: [prealert_at, acknowledgment_due_at, initial_due_at, announced_at, days_remaining, final_response_at]
          properties:
            prealert_at: { $ref: '#/components/schemas/NullableDateTime' }
            acknowledgment_due_at: { $ref: '#/components/schemas/NullableDateTime' }
            initial_due_at: { $ref: '#/components/schemas/NullableDateTime' }
            announced_at: { $ref: '#/components/schemas/NullableDateTime' }
            days_remaining: { type: [integer, 'null'], description: Jours restants (négatif si dépassé) ; null si répondu. }
            final_response_at: { $ref: '#/components/schemas/NullableDateTime' }
    DeadlinesPage:
      type: object
      required: [data, meta, summary, rules, rule_is_demo]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/DeadlineQueueItem' }
        meta: { $ref: '#/components/schemas/PaginationMeta' }
        summary:
          type: object
          required: [a_risque, en_retard, clos_en_retard]
          properties:
            a_risque: { type: integer }
            en_retard: { type: integer }
            clos_en_retard: { type: integer }
        rules:
          type: array
          description: Règles de délai actives.
          items: { $ref: '#/components/schemas/DeadlineRule' }
        rule_is_demo: { type: boolean }
...
    DeadlineRule:
      type: object
      description: Règle de délai versionnée (DeadlineRuleResource). Une règle validée est immuable.
      required: [id, code, label, kind, kind_label, unit, unit_label, duration, start_point, source_type, source_type_label, source_reference, effective_from, effective_to, version, status, status_label, is_demo, validated_by, validated_at, notes]
      properties:
        id: { type: integer }
        code: { type: string, example: REP_FINALE }
        label: { type: string }
        kind: { $ref: '#/components/schemas/DeadlineKind' }
        kind_label: { type: string }
        unit: { $ref: '#/components/schemas/DeadlineUnit' }
        unit_label: { type: string }
        duration: { type: integer, minimum: 0, maximum: 365 }
        start_point: { type: string, example: received_at }
        source_type: { $ref: '#/components/schemas/RuleSourceType' }
        source_type_label: { type: string }
        source_reference: { type: string }
        effective_from: { type: [string, 'null'], format: date }
        effective_to: { type: [string, 'null'], format: date }
        version: { type: integer }
        status: { $ref: '#/components/schemas/RuleStatus' }
        status_label: { type: string }
        is_demo: { type: boolean }
        validated_by:
          oneOf:
            - $ref: '#/components/schemas/IdName'
            - type: 'null'
        validated_at: { $ref: '#/components/schemas/NullableDateTime' }
        notes: { type: [string, 'null'] }
...